LayoutTests/fast/forms/date/date-editable-components/date-picker-show-on-space-keypress-expected.txt

Tests that a date input's calendar view can be opened using the space key, without any mouse interaction.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS showingPicker is false
Focusing input using tab key and then pressing space key.
PASS showingPicker is true
PASS successfullyParsed is true

TEST COMPLETE
